Extract an ErbRenderer class from OutcomePresenter
I want to re-use this for the StartNodePresenter.
